refactor: simplify the Unity Editor package
