+++ editor/Editor/MiniJson.cs
@@ -27,55 +27,15 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 // Bundled into PrefabLens from: https://gist.github.com/darktable/1411710 (rev 513f1c0, MIT)
-// Change: namespace MiniJSON -> PrefabLens.MiniJson only.
+// Changes: namespace MiniJSON -> PrefabLens.MiniJson; removed unused serialization and its example.
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
 namespace PrefabLens.MiniJson
 {
-    // Example usage:
-    //
-    //  using UnityEngine;
-    //  using System.Collections;
-    //  using System.Collections.Generic;
-    //  using MiniJSON;
-    //
-    //  public class MiniJSONTest : MonoBehaviour {
-    //      void Start () {
-    //          var jsonString = "{ \"array\": [1.44,2,3], " +
-    //                          "\"object\": {\"key1\":\"value1\", \"key2\":256}, " +
-    //                          "\"string\": \"The quick brown fox \\\"jumps\\\" over the lazy dog \", " +
-    //                          "\"unicode\": \"\\u3041 Men\u00fa sesi\u00f3n\", " +
-    //                          "\"int\": 65536, " +
-    //                          "\"float\": 3.1415926, " +
-    //                          "\"bool\": true, " +
-    //                          "\"null\": null }";
-    //
-    //          var dict = Json.Deserialize(jsonString) as Dictionary<string,object>;
-    //
-    //          Debug.Log("deserialized: " + dict.GetType());
-    //          Debug.Log("dict['array'][0]: " + ((List<object>) dict["array"])[0]);
-    //          Debug.Log("dict['string']: " + (string) dict["string"]);
-    //          Debug.Log("dict['float']: " + (double) dict["float"]); // floats come out as doubles
-    //          Debug.Log("dict['int']: " + (long) dict["int"]); // ints come out as longs
-    //          Debug.Log("dict['unicode']: " + (string) dict["unicode"]);
-    //
-    //          var str = Json.Serialize(dict);
-    //
-    //          Debug.Log("serialized: " + str);
-    //      }
-    //  }
-
-    /// <summary>
-    /// This class encodes and decodes JSON strings.
-    /// Spec. details, see http://www.json.org/
-    ///
-    /// JSON uses Arrays and Objects. These correspond here to the datatypes IList and IDictionary.
-    /// All numbers are parsed to doubles.
-    /// </summary>
+    /// Reads CLI JSON into lists, dictionaries, and scalar values.
     public static class Json
     {
         /// <summary>
@@ -85,7 +45,6 @@ namespace PrefabLens.MiniJson
         /// <returns>An List&lt;object&gt;, a Dictionary&lt;string, object&gt;, a double, an integer,a string, null, true, or false</returns>
         public static object Deserialize(string json)
         {
-            // save the string for debug information
             if (json == null)
             {
                 return null;
@@ -431,196 +390,6 @@ namespace PrefabLens.MiniJson
                     }
 
                     return TOKEN.NONE;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Converts a IDictionary / IList object or a simple type (string, int, etc.) into a JSON string
-        /// </summary>
-        /// <param name="json">A Dictionary&lt;string, object&gt; / List&lt;object&gt;</param>
-        /// <returns>A JSON encoded string, or null if object 'json' is not serializable</returns>
-        public static string Serialize(object obj)
-        {
-            return Serializer.Serialize(obj);
-        }
-
-        sealed class Serializer
-        {
-            StringBuilder builder;
-
-            Serializer()
-            {
-                builder = new StringBuilder();
-            }
-
-            public static string Serialize(object obj)
-            {
-                var instance = new Serializer();
-
-                instance.SerializeValue(obj);
-
-                return instance.builder.ToString();
-            }
-
-            void SerializeValue(object value)
-            {
-                IList asList;
-                IDictionary asDict;
-                string asStr;
-
-                if (value == null)
-                {
-                    builder.Append("null");
-                }
-                else if ((asStr = value as string) != null)
-                {
-                    SerializeString(asStr);
-                }
-                else if (value is bool)
-                {
-                    builder.Append((bool)value ? "true" : "false");
-                }
-                else if ((asList = value as IList) != null)
-                {
-                    SerializeArray(asList);
-                }
-                else if ((asDict = value as IDictionary) != null)
-                {
-                    SerializeObject(asDict);
-                }
-                else if (value is char)
-                {
-                    SerializeString(new string((char)value, 1));
-                }
-                else
-                {
-                    SerializeOther(value);
-                }
-            }
-
-            void SerializeObject(IDictionary obj)
-            {
-                bool first = true;
-
-                builder.Append('{');
-
-                foreach (object e in obj.Keys)
-                {
-                    if (!first)
-                    {
-                        builder.Append(',');
-                    }
-
-                    SerializeString(e.ToString());
-                    builder.Append(':');
-
-                    SerializeValue(obj[e]);
-
-                    first = false;
-                }
-
-                builder.Append('}');
-            }
-
-            void SerializeArray(IList anArray)
-            {
-                builder.Append('[');
-
-                bool first = true;
-
-                foreach (object obj in anArray)
-                {
-                    if (!first)
-                    {
-                        builder.Append(',');
-                    }
-
-                    SerializeValue(obj);
-
-                    first = false;
-                }
-
-                builder.Append(']');
-            }
-
-            void SerializeString(string str)
-            {
-                builder.Append('\"');
-
-                char[] charArray = str.ToCharArray();
-                foreach (var c in charArray)
-                {
-                    switch (c)
-                    {
-                        case '"':
-                            builder.Append("\\\"");
-                            break;
-                        case '\\':
-                            builder.Append("\\\\");
-                            break;
-                        case '\b':
-                            builder.Append("\\b");
-                            break;
-                        case '\f':
-                            builder.Append("\\f");
-                            break;
-                        case '\n':
-                            builder.Append("\\n");
-                            break;
-                        case '\r':
-                            builder.Append("\\r");
-                            break;
-                        case '\t':
-                            builder.Append("\\t");
-                            break;
-                        default:
-                            int codepoint = Convert.ToInt32(c);
-                            if ((codepoint >= 32) && (codepoint <= 126))
-                            {
-                                builder.Append(c);
-                            }
-                            else
-                            {
-                                builder.Append("\\u");
-                                builder.Append(codepoint.ToString("x4"));
-                            }
-                            break;
-                    }
-                }
-
-                builder.Append('\"');
-            }
-
-            void SerializeOther(object value)
-            {
-                // NOTE: decimals lose precision during serialization.
-                // They always have, I'm just letting you know.
-                // Previously floats and doubles lost precision too.
-                if (value is float)
-                {
-                    builder.Append(((float)value).ToString("R"));
-                }
-                else if (
-                    value is int
-                    || value is uint
-                    || value is long
-                    || value is sbyte
-                    || value is byte
-                    || value is short
-                    || value is ushort
-                    || value is ulong
-                )
-                {
-                    builder.Append(value);
-                }
-                else if (value is double || value is decimal)
-                {
-                    builder.Append(Convert.ToDouble(value).ToString("R"));
-                }
-                else
-                {
-                    SerializeString(value.ToString());
                 }
             }
         }
