+++ editor/Tests/Editor/CliTests.Download.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73d398c3777541c8a0ba0f329813f31a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
