+++ editor/Tests/Editor/CliTests.Run.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbd2e06e8bdf4f5bb63214551a722fe6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
